+++ packages/analytics/analytics-utilities/src/types/query-bridge.ts
@@ -21,18 +21,18 @@ export type DatasourceAwareQuery = BasicDatasourceQuery | AdvancedDatasourceQuer
 
 export interface AnalyticsBridge {
   // Issue queries to the KAnalytics API
-  queryFn: (query: DatasourceAwareQuery, abortController: AbortController) => Promise<ExploreResultV4>,
+  queryFn: (query: DatasourceAwareQuery, abortController: AbortController) => Promise<ExploreResultV4>
 
   // Determine the current org's analytics config
-  configFn: () => Promise<AnalyticsConfigV2>,
+  configFn: () => Promise<AnalyticsConfigV2>
 
   // Evaluate feature flags (if applicable)
-  evaluateFeatureFlagFn: <T = boolean>(key: string, defaultValue: T) => T,
+  evaluateFeatureFlagFn: <T = boolean>(key: string, defaultValue: T) => T
 
   // Define the location of explore to enable jump-to-explore.
   // Async because there might need to be permissions checks.
-  exploreBaseUrl?: () => Promise<string>,
+  exploreBaseUrl?: () => Promise<string>
 
   // Dynamically provide certain components that aren't available in all environments
-  fetchComponent?: (name: string) => Promise<Component>,
+  fetchComponent?: (name: string) => Promise<Component>
 }