+++ packages/analytics/dashboard-renderer/src/components/BaseAnalyticsChartRenderer.vue
@@ -12,7 +12,6 @@
         :allow-csv-export="chartOptions.allowCsvExport"
         :chart-data="data"
         :chart-options="options"
-        :chart-title="!hasKebabMenuAccess && chartOptions.chartTitle || ''"
         legend-position="bottom"
         :show-menu="context.editable"
         :synthetics-data-key="chartOptions.syntheticsDataKey"
@@ -51,8 +50,6 @@ const emit = defineEmits<{
 }>()
 const { i18n } = composables.useI18n()
 const { evaluateFeatureFlag } = composables.useEvaluateFeatureFlag()
-
-const hasKebabMenuAccess = evaluateFeatureFlag('ma-3043-analytics-chart-kebab-menu', false)
 const hasFinegrainedAbsoluteTimerangeAccess = evaluateFeatureFlag('explore-v4-fine-granularity', false)
 
 const options = computed((): AnalyticsChartOptions => ({