+++ packages/entities/entities-plugins/src/components/free-form/shared/utils.ts
@@ -1,5 +1,5 @@
 export function toSelectItems<T extends string>(
   items: T[],
-): { value: T; label: T }[] {
+): { value: T, label: T }[] {
   return items.map((item) => ({ value: item, label: item }))
 }