+++ packages/entities/entities-plugins/src/types/plugins/shared.d.ts
@@ -4,11 +4,11 @@ interface Field {
   type: string
   values?: string[]
   id?: string
-  default?: string,
-  required?: boolean,
-  placeholder?: string,
-  hint?: string,
-  help?: string,
+  default?: string
+  required?: boolean
+  placeholder?: string
+  hint?: string
+  help?: string
   inputType?: 'text' | 'number'
 }
 
@@ -20,7 +20,7 @@ interface ArrayItem extends Field {
     class?: string
     style?: {
       minWidth?: string
-    },
+    }
     type?: string
     rows?: number
     max?: boolean | number
@@ -28,19 +28,19 @@ interface ArrayItem extends Field {
 }
 
 export interface ItemsSchema {
-  type: string,
-  default?: object,
+  type: string
+  default?: object
   schema: {
     fields: Array<Field | ArrayItem>
   }
 }
 
 export interface PluginBasicSchema {
-  title: string,
-  plugin: string,
-  name: string,
-  endpoint: string,
-  schemaEndpoint: string,
+  title: string
+  plugin: string
+  name: string
+  endpoint: string
+  schemaEndpoint: string
 }
 
 export interface CommonSchemaFields {