+++ packages/entities/entities-vaults/src/components/VaultConfigCard.vue
@@ -31,8 +31,8 @@ import composables from '../composables'
 
 defineEmits<{
   (e: 'loading', isLoading: boolean): void
-  (e: 'fetch:error', error: AxiosError): void,
-  (e: 'fetch:success', data: Record<string, any>): void,
+  (e: 'fetch:error', error: AxiosError): void
+  (e: 'fetch:success', data: Record<string, any>): void
 }>()
 
 // Component props - This structure must exist in ALL entity components, with the exclusion of unneeded action props (e.g. if you don't need `canDelete`, just exclude it)