+++ packages/entities/entities-vaults/src/components/VaultForm.vue
@@ -608,9 +608,9 @@ const props = defineProps({
 })
 
 const emit = defineEmits<{
-  (e: 'update', data: VaultStateFields): void,
-  (e: 'error', error: AxiosError): void,
-  (e: 'loading', isLoading: boolean): void,
+  (e: 'update', data: VaultStateFields): void
+  (e: 'error', error: AxiosError): void
+  (e: 'loading', isLoading: boolean): void
 }>()
 
 const { i18nT, i18n: { t } } = composables.useI18n()