+++ packages/analytics/analytics-utilities/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.3.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-utilities@9.3.3...@kong-ui-public/analytics-utilities@9.3.4) (2025-05-05)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-utilities
+
+
+
+
+
+## [9.3.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-utilities@9.3.2...@kong-ui-public/analytics-utilities@9.3.3) (2025-05-02)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-utilities
+
+
+
+
+
 ## [9.3.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-utilities@9.3.1...@kong-ui-public/analytics-utilities@9.3.2) (2025-04-24)