+++ packages/core/app-layout/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.4.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/app-layout@4.4.2...@kong-ui-public/app-layout@4.4.3) (2025-05-05)
+
+**Note:** Version bump only for package @kong-ui-public/app-layout
+
+
+
+
+
+## [4.4.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/app-layout@4.4.1...@kong-ui-public/app-layout@4.4.2) (2025-05-02)
+
+**Note:** Version bump only for package @kong-ui-public/app-layout
+
+
+
+
+
+## [4.4.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/app-layout@4.4.0...@kong-ui-public/app-layout@4.4.1) (2025-05-01)
+
+
+### Bug Fixes
+
+* **deps:** remove recursive kongponents dependency ([#2100](https://github.com/Kong/public-ui-components/issues/2100)) ([3006571](https://github.com/Kong/public-ui-components/commit/3006571fc6e0c522a622bfed81ca616b5e072555))
+
+
+
+
+
 # [4.4.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/app-layout@4.3.17...@kong-ui-public/app-layout@4.4.0) (2025-04-25)