+++ packages/core/expressions/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.32](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/expressions@0.9.31...@kong-ui-public/expressions@0.9.32) (2025-05-05)
+
+**Note:** Version bump only for package @kong-ui-public/expressions
+
+
+
+
+
+## [0.9.31](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/expressions@0.9.30...@kong-ui-public/expressions@0.9.31) (2025-05-02)
+
+**Note:** Version bump only for package @kong-ui-public/expressions
+
+
+
+
+
+## [0.9.30](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/expressions@0.9.29...@kong-ui-public/expressions@0.9.30) (2025-05-01)
+
+
+### Bug Fixes
+
+* **deps:** remove recursive kongponents dependency ([#2100](https://github.com/Kong/public-ui-components/issues/2100)) ([3006571](https://github.com/Kong/public-ui-components/commit/3006571fc6e0c522a622bfed81ca616b5e072555))
+
+
+
+
+
 ## [0.9.29](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/expressions@0.9.28...@kong-ui-public/expressions@0.9.29) (2025-04-29)
 
 **Note:** Version bump only for package @kong-ui-public/expressions