+++ packages/entities/entities-plugins/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.18.14](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.18.13...@kong-ui-public/entities-plugins@9.18.14) (2025-05-05)
+
+**Note:** Version bump only for package @kong-ui-public/entities-plugins
+
+
+
+
+
+## [9.18.13](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.18.12...@kong-ui-public/entities-plugins@9.18.13) (2025-05-02)
+
+**Note:** Version bump only for package @kong-ui-public/entities-plugins
+
+
+
+
+
+## [9.18.12](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.18.11...@kong-ui-public/entities-plugins@9.18.12) (2025-05-01)
+
+
+### Bug Fixes
+
+* **deps:** remove recursive kongponents dependency ([#2100](https://github.com/Kong/public-ui-components/issues/2100)) ([3006571](https://github.com/Kong/public-ui-components/commit/3006571fc6e0c522a622bfed81ca616b5e072555))
+
+
+
+
+
 ## [9.18.11](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.18.10...@kong-ui-public/entities-plugins@9.18.11) (2025-04-29)
 
 **Note:** Version bump only for package @kong-ui-public/entities-plugins